moving generator to inflect and some refactorings
