+++ Gopkg.lock
@@ -2,6 +2,12 @@
 
 
 [[projects]]
+  branch = "master"
+  name = "github.com/cockroachdb/cockroach-go"
+  packages = ["crdb"]
+  revision = "0d8b4682f140f0fe486ef7e3d2f70665f3066906"
+
+[[projects]]
   name = "github.com/davecgh/go-spew"
   packages = ["spew"]
   revision = "346938d642f2ec3594ed81d874461961cd0faa76"
@@ -235,6 +241,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "3d1ffac3adeb9084a7027b21116afa6d130b18afd2b5a176bf432307db8cb8cc"
+  inputs-digest = "c3cc22c0928de145c466255410ccd1243a0c5743b1fba8acedb6967e9e0aa55f"
   solver-name = "gps-cdcl"
   solver-version = 1
